# sdks/sdkwork-models-backend-sdk/openapi/sdkwork-models-backend-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Models Backend API",
    "version": "0.1.0",
    "description": "Intelligence catalog API owned by sdkwork-models. Materialized from composed host catalog mount OpenAPI authority."
  },
  "servers": [
    {
      "url": "/backend/v3/api"
    }
  ],
  "paths": {
...
    "/backend/v3/api/ai/model_rankings": {
      "get": {
        "description": "List. Reads ai. Writes none. File targets none.",
        "operationId": "modelRankings.list",
        "parameters": [
          {
            "description": "Ranking snapshot scope.",
            "in": "query",
            "name": "rank_scope",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Vendor code filter.",
            "in": "query",
            "name": "vendor_code",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Model modality filter.",
            "in": "query",
            "name": "modality",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Free-text ranking search query.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Number of ranking rows returned in the current page.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 200,
              "format": "int32",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRankingsListResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Server Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Error response."
          }
        },
        "security": [
          {
            "AccessToken": [],
            "AuthToken": []
          }
        ],
        "summary": "List",
        "tags": [
          "ai"
        ],
        "x-contract-kind": "read",
        "x-file-targets": [],
        "x-read-sources": [
          "ai"
        ],
        "x-route-scope": "admin",
        "x-sdk-domain": "intelligence",
        "x-sdkwork-api-authority": "sdkwork-models-backend-api",
        "x-sdkwork-domain": "intelligence",
        "x-sdkwork-owner": "sdkwork-models",
        "x-sdkwork-resource": "modelRankings",
        "x-sdkwork-source-route-crate": "sdkwork-routes-models-catalog-backend-api",
        "x-source-file": "tools/bootstrap_frontend_contract_from_route_manifest.py",
        "x-write-tables": []
      }
    },
...
    "/backend/v3/api/ai/model_rankings/refresh": {
      "post": {
        "description": "Refresh. Reads ai. Writes ops_audit_log. File targets none.",
        "operationId": "modelRankings.refresh",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRankingsRefreshResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Server Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Error response."
          }
        },
        "security": [
          {
            "AccessToken": [],
            "AuthToken": []
          }
        ],
        "summary": "Refresh",
        "tags": [
          "ai"
        ],
        "x-contract-kind": "create",
        "x-file-targets": [],
        "x-read-sources": [
          "ai"
        ],
        "x-route-scope": "admin",
        "x-sdk-domain": "intelligence",
        "x-sdkwork-api-authority": "sdkwork-models-backend-api",
        "x-sdkwork-domain": "intelligence",
        "x-sdkwork-owner": "sdkwork-models",
        "x-sdkwork-resource": "modelRankings",
        "x-sdkwork-source-route-crate": "sdkwork-routes-models-catalog-backend-api",
        "x-source-file": "tools/bootstrap_frontend_contract_from_route_manifest.py",
        "x-write-tables": [
          "ops_audit_log"
        ]
      }
    },
...
    "/backend/v3/api/ai/models/refresh": {
      "post": {
        "description": "Refresh. Reads ai. Writes ops_audit_log. File targets none.",
        "operationId": "models.refresh",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/ModelsRefreshResult"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Server Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Error response."
          }
        },
        "security": [
          {
            "AccessToken": [],
            "AuthToken": []
          }
        ],
        "summary": "Refresh",
        "tags": [
          "ai"
        ],
        "x-contract-kind": "create",
        "x-file-targets": [],
        "x-read-sources": [
          "ai"
        ],
        "x-route-scope": "admin",
        "x-sdk-domain": "intelligence",
        "x-sdkwork-api-authority": "sdkwork-models-backend-api",
        "x-sdkwork-domain": "intelligence",
        "x-sdkwork-owner": "sdkwork-models",
        "x-sdkwork-resource": "models",
        "x-sdkwork-source-route-crate": "sdkwork-routes-models-catalog-backend-api",
        "x-source-file": "tools/bootstrap_frontend_contract_from_route_manifest.py",
        "x-write-tables": [
          "ops_audit_log"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "in": "header",
        "name": "Access-Token",
        "type": "apiKey"
      },
      "AuthToken": {
        "bearerFormat": "SDKWork auth token",
        "scheme": "bearer",
        "type": "http"
      }
    },
    "schemas": {
...
      "ModelRankingsListResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "additionalProperties": false,
            "description": "Object schema used by model rankings list result.",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ModelRankingsPage"
                  }
                ],
                "description": "Data field on model rankings list result."
              }
            },
            "required": [
              "data"
            ],
            "type": "object"
          }
        ],
        "description": "Model rankings list result schema exposed by Claw Router.",
        "x-operation-id": "modelRankings.list"
      },
      "ModelRankingsPage": {
        "additionalProperties": false,
        "description": "Model rankings page schema exposed by Claw Router.",
        "name": "ModelRankingsPage",
        "properties": {
          "history": {
            "description": "History field on model rankings page.",
            "items": {
              "additionalProperties": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JsonValue"
                  }
                ],
                "description": "History field on model rankings page."
              },
              "description": "History field on model rankings page.",
              "type": "object"
            },
            "type": "array"
          },
          "items": {
            "description": "Items field on model rankings page.",
            "items": {
              "additionalProperties": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JsonValue"
                  }
                ],
                "description": "Items field on model rankings page."
              },
              "description": "Items field on model rankings page.",
              "type": "object"
            },
            "type": "array"
          },
          "pageInfo": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PageInfo"
              }
            ],
            "description": "Page info field on model rankings page."
          },
          "source": {
            "additionalProperties": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/JsonValue"
                }
              ],
              "description": "Source field on model rankings page."
            },
            "description": "Source field on model rankings page.",
            "type": "object"
          }
        },
        "required": [
          "items",
          "history",
          "source",
          "pageInfo"
        ],
        "type": "object"
      },
      "ModelRankingsRefreshResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "additionalProperties": false,
            "description": "Object schema used by model rankings refresh result.",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/NoData"
                  }
                ],
                "description": "No business data returned by this operation."
              }
            },
            "required": [
              "data"
            ],
            "type": "object"
          }
        ],
        "description": "Model rankings refresh result schema exposed by Claw Router.",
        "x-operation-id": "modelRankings.refresh"
      },
...
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "i18nKey": {
            "type": "string",
            "description": "Optional stable localization key such as errors.result.40001."
          },
          "locale": {
            "type": "string",
            "description": "Optional effective BCP 47 locale used by framework message mapping."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
    }
  }
}